This should do the trick [untested]
If it works, would be good to add a UI control to enable/disable this check along with a warning about the server's resolvers needing to not include localhost for it to be effective.

+++ plib/scripts/slave-dns.php
@@ -23,16 +23,20 @@ foreach ($data as $task) {
     }
 
     $rndc = new Modules_SlaveDnsManager_Rndc();
+    
+    if ($rndc->isAuthoritative($domain){
 
-    switch ($command) {
-        case 'create':
-            $rndc->addZone($domain);
-            break;
-        case 'update':
-            $rndc->updateZone($domain);
-            break;
-        case 'delete':
-            $rndc->deleteZone($domain);
-            break;
+      switch ($command) {
+          case 'create':
+              $rndc->addZone($domain);
+              break;
+          case 'update':
+              $rndc->updateZone($domain);
+              break;
+          case 'delete':
+              $rndc->deleteZone($domain);
+              break;
+      }
+    
     }
 }
